+++ resources/views/livewire-radial-chart.blade.php
@@ -1,0 +1,7 @@
+<div
+    style="width: 100%; height: 100%;"
+    x-data="{ ...livewireChartsRadialChart() }"
+    x-init="init()"
+>
+    <div wire:ignore x-ref="container"></div>
+</div>