+++ src/Facades/LivewireCharts.php
@@ -7,6 +7,7 @@ use Asantibanez\LivewireCharts\Models\ColumnChartModel;
 use Asantibanez\LivewireCharts\Models\LineChartModel;
 use Asantibanez\LivewireCharts\Models\PieChartModel;
 use Asantibanez\LivewireCharts\Models\RadarChartModel;
+use Asantibanez\LivewireCharts\Models\RadialChartModel;
 use Asantibanez\LivewireCharts\Models\TreeMapChartModel;
 use Illuminate\Support\Facades\Facade;
 
@@ -21,6 +22,7 @@ use Illuminate\Support\Facades\Facade;
  * @method static PieChartModel pieChartModel()
  * @method static RadarChartModel radarChartModel()
  * @method static TreeMapChartModel treeMapChartModel()
+ * @method static RadialChartModel radialChartModel()
  */
 class LivewireCharts extends Facade
 {