+++ src/LivewireChartsServiceProvider.php
@@ -7,6 +7,7 @@ use Asantibanez\LivewireCharts\Charts\LivewireColumnChart;
 use Asantibanez\LivewireCharts\Charts\LivewireLineChart;
 use Asantibanez\LivewireCharts\Charts\LivewirePieChart;
 use Asantibanez\LivewireCharts\Charts\LivewireRadarChart;
+use Asantibanez\LivewireCharts\Charts\LivewireRadialChart;
 use Asantibanez\LivewireCharts\Charts\LivewireTreeMapChart;
 use Asantibanez\LivewireCharts\Console\InstallCommand;
 use Illuminate\Support\Facades\Blade;
@@ -73,6 +74,7 @@ class LivewireChartsServiceProvider extends ServiceProvider
         Livewire::component('livewire-area-chart', LivewireAreaChart::class);
         Livewire::component('livewire-radar-chart', LivewireRadarChart::class);
         Livewire::component('livewire-tree-map-chart', LivewireTreeMapChart::class);
+        Livewire::component('livewire-radial-chart', LivewireRadialChart::class);
     }
 
     private function registerDirectives()